net: mscc: ocelot: dynamically allocate ptp_rule
This fixes a NULL pointer bug on driver unbind:

[  275.753961] Unable to handle kernel paging request at virtual address 00000174a8a73848
[  275.917700] Call trace:
[  275.920148]  kfree+0x48/0xc8
[  275.923033]  ocelot_ace_rule_offload_del+0x10c/0x1f0
[  275.928011]  ocelot_deinit+0xa0/0xa8
[  275.931594]  felix_teardown+0x2c/0x38
[  275.935264]  dsa_switch_teardown+0x44/0x68
[  275.939369]  dsa_tree_teardown_switches+0x5c/0x80
[  275.944084]  dsa_unregister_switch+0x70/0x90
[  275.948364]  felix_pci_remove+0x1c/0x30
[  275.952210]  pci_device_remove+0x3c/0xe0

which happens because ocelot_ace_rule_del() tries to kfree() memory that
hasn't been dynamically allocated. Fix that by allocating it
dynamically, so that the API is respected.

In the general case, we don't have any real use case for
statically-defined VCAP IS2 rules, so dynamic allocation for this
one-off case is fine - even this ptp_rule got pushback upstream and we
need to find a different mechanism there to ensure 1588 frames are
trapped to the CPU (which is likely going to be through multicast
filtering).

Reported-by: Claudiu Manoil <claudiu.manoil@nxp.com>
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Yangbo Lu <yangbo.lu@nxp.com>