net: mscc: ocelot: fix unregistering the PTP clock twice
When merged tag 'v5.4.47' into lf-5.4.y, the changes merging of below two
commits were not handled properly. It is causing PTP clock unregiseter is
executed twice. One is in ocelot_deinit_timestamp(), and the other one is
in ocelot_deinit().

9dc3383 net: mscc: ocelot: unregister the PTP clock on deinit
157560f net: mscc: ocelot: unregister the PTP clock on deinit

This patch is to remove the duplicated ptp_clock_unregister().

Signed-off-by: Yangbo Lu <yangbo.lu@nxp.com>